add Context.LookupRequestValue
It is a common usecase for us to pull request values out of the request
path match in the HTTP router. With gorilla/mux, this works fine because
mux.Vars(r) gives a map[string]string, i.e. just what goslo.policy
wants.

But the new HTTP router functionality in the standard library does not
have that: You can only query for one path value at a time. This PR adds
API to make an integration with the net/http router possible.